# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a toge server API documentation.",
        "title": "toge API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/timezone/convert": {
            "get": {
                "description": "将时间从一个时区转换到另一个时区",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "时区"
                ],
                "summary": "转换时区",
                "parameters": [
                    {
                        "type": "string",
                        "example": "2023-01-01 12:00:00",
                        "description": "时间字符串",
                        "name": "time",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "Asia/Shanghai",
                        "description": "源时区",
                        "name": "from_timezone",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "America/New_York",
                        "description": "目标时区",
                        "name": "to_timezone",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/git_lulumia_fun_root_toge-api_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/git_lulumia_fun_root_toge-api_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
        "/timezone/format": {
            "get": {
                "description": "将时间戳格式化为指定时区的时间字符串",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "时区"
                ],
                "summary": "格式化时间",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "example": 1672531200,
                        "description": "时间戳",
                        "name": "timestamp",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "Asia/Shanghai",
                        "description": "时区名称",
                        "name": "timezone",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "2006-01-02 15:04:05",
                        "description": "时间格式",
                        "name": "format",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/git_lulumia_fun_root_toge-api_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/git_lulumia_fun_root_toge-api_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/timezone/parse": {
            "get": {
                "description": "在指定时区解析时间字符串",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "时区"
                ],
                "summary": "解析时间字符串",
                "parameters": [
                    {
                        "type": "string",
                        "example": "2023-01-01 12:00:00",
                        "description": "时间字符串",
                        "name": "time",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "Asia/Shanghai",
                        "description": "时区名称",
                        "name": "timezone",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/git_lulumia_fun_root_toge-api_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/git_lulumia_fun_root_toge-api_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/timezone/time": {
            "get": {
                "description": "获取指定时区的当前时间",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "时区"
                ],
                "summary": "获取指定时区的时间",
                "parameters": [
                    {
                        "type": "string",
                        "example": "Asia/Shanghai",
                        "description": "时区名称",
                        "name": "timezone",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/git_lulumia_fun_root_toge-api_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/git_lulumia_fun_root_toge-api_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "git_lulumia_fun_root_toge-api_pkg_response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据"
                },
                "error": {
                    "description": "错误信息",
                    "type": "string"
                },
                "message": {
                    "description": "响应消息",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}